Правки в парсинге

+++ database/migrations/2022_06_26_131002_create_price_reductions_table.php
@@ -22,7 +22,7 @@ class CreatePriceReductionsTable extends Migration
             $table->dateTime('end_time')->nullable();
             $table->decimal('price', 15,2)->default(0.00);
             $table->decimal('deposit', 15,2)->default(0.00);
-            $table->integer('percent')->default(0);
+            $table->bigInteger('percent')->default(0);
             $table->boolean('is_system')->default(false);
             $table->timestamps();
         });
